requirements.txt
requests>=2.28.0
PyYAML>=6.0
unidiff>=0.7.4
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
google-generativeai>=0.3.0
bandit>=1.7.0
flake8>=6.0.0
GitPython>=3.1.0